Create plugin for Scala publishing conventions
