--- buildSrc/src/main/kotlin/xef-scala-documentation.gradle.kts
@@ -1,1 +1,0 @@
-apply<ScalaDocumentationPlugin>()
