+++ buildSrc/src/main/kotlin/xef-scala-publishing-conventions.gradle.kts
@@ -1,0 +1,1 @@
+apply<ScalaPublishingConventionsPlugin>()
